requirements.txt
Django>=3
Pillow>=7.1.0
PyMySQL==0.9.3
pytz==2020.1
sqlparse==0.3.1
django-environ
pyjwt
pyunitreport
selenium